modification du text

+++ SITIENS/ViewModel/UserSettingsViewModel.swift
@@ -93,13 +93,13 @@ class UserSettingsViewModel{
         }
     }
     
-    func showNumberOfGlass(chooseBottle:String, name:String) -> Int {
+    func showNumberOfGlass(chooseBottle:String, name:String) -> Double {
         let waterQuantity = quantityWaterNumber(chooseBottle:chooseBottle)
         let updatedProfile = updateWater(type :name)
         
         let totalMl = updatedProfile * 1000
         print("totalMl : \(totalMl)")
-        return Int(totalMl / CGFloat(waterQuantity))
+        return totalMl / CGFloat(waterQuantity)
     }
     
     func chooseBottleOfWater(name type: String) -> String {
